Update KANClassifier.py functions
get_params() was uptaded to include other parameters
related to the models training. in fit(), the model
only declares its fitted after the training is done.
Added state_dict() and load_state_dict(). Theres
also some changes in fit() related to the input,
but for now most of it is commented.